Re-organize Address services to be use outside of Account context
This way we can re-use them for guest users on Checkout and potentially in Platform API.

Left the old classes / dependencies for past versions compatiblity
