+++ core/lib/spree/core/app_dependencies.rb
@@ -16,7 +16,7 @@ module Spree
       :classification_reposition_service, :credit_cards_destroy_service, :cart_associate_service, :cart_change_currency_service,
       :line_item_create_service, :line_item_update_service, :line_item_destroy_service,
       :order_approve_service, :order_cancel_service, :shipment_change_state_service, :shipment_update_service,
-      :payment_create_service
+      :payment_create_service, :address_create_service, :address_update_service
     ].freeze
 
     attr_accessor *INJECTION_POINTS
@@ -81,8 +81,10 @@ module Spree
       # account
       @account_create_service = 'Spree::Account::Create'
       @account_update_service = 'Spree::Account::Update'
-      @account_create_address_service = 'Spree::Account::Addresses::Create'
-      @account_update_address_service = 'Spree::Account::Addresses::Update'
+
+      # addresses
+      @address_create_service = 'Spree::Addresses::Create'
+      @address_update_service = 'Spree::Addresses::Update'
 
       # credit cards
       @credit_cards_destroy_service = 'Spree::CreditCards::Destroy'
